# static/openapi.yaml
openapi: "3.0.3"
info:
  title: AiPayGen API
  description: |
    AI-powered API platform with 244+ tools. Pay-per-call with API keys (Bearer apk_xxx),
    x402 V2 USDC micropayments (multi-chain: Base, Solana, Stellar), or use the free tier (1 call/day per IP).
    Supports autonomous agent micropayments — no API key or signup required.

    Base URL: https://api.aipaygen.com
  version: "1.8.1"
  contact:
    email: hello@aipaygen.com
    url: https://aipaygen.com
  license:
    name: Proprietary
servers:
  - url: https://api.aipaygen.com
    description: Production
security:
  - BearerApiKey: []
  - AgentJWT: []
  - {}
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: "API key with apk_ prefix. Obtain via POST /auth/generate-key or POST /credits/buy."
    AgentJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Agent JWT obtained via /agents/challenge + /agents/verify wallet auth flow."
  schemas:
...
    AiToolResponse:
      type: object
      properties:
        result:
          type: string
        model:
          type: string
        _meta:
          type: object
...
paths:
...
  /code:
    post:
      tags: [AI Tools]
      summary: Generate code
      description: Generate code in any language from a natural language description.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: "A function that sorts a list of dictionaries by key"
                language:
                  type: string
                  default: Python
                model:
                  type: string
                  description: "Model to use (e.g. claude-haiku, gpt-4o, deepseek)"
      responses:
        "200":
          description: Generated code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiToolResponse"
        "400":
          description: Missing description
...
  /email:
    post:
      tags: [AI Tools]
      summary: Generate professional emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                tone:
                  type: string
                  default: professional
                model:
                  type: string
      responses:
        "200":
          description: Generated email
...